+++ src/runtime/vdom/vdom-render.ts
@@ -16,7 +16,7 @@ import { h, isHost, newVNode } from './h';
 import { updateElement } from './update-element';
 
 let scopeId: string;
-let contentRef: d.RenderNode;
+let contentRef: d.RenderNode | undefined;
 let hostTagName: string;
 let useNativeShadowDom = false;
 let checkSlotFallbackVisibility = false;
@@ -143,9 +143,6 @@ const createElm = (oldParentVNode: d.VNode, newParentVNode: d.VNode, childIndex:
       // remember the content reference comment
       elm['s-sr'] = true;
 
-      // Persist the name of the slot that this slot was going to be projected into.
-      elm['s-fs'] = newVNode.$attrs$?.slot;
-
       // remember the content reference comment
       elm['s-cr'] = contentRef;
 
@@ -185,19 +182,18 @@ const relocateToHostRoot = (parentElm: Element) => {
 
   const host = parentElm.closest(hostTagName.toLowerCase());
   if (host != null) {
-    for (const childNode of Array.from(parentElm.childNodes) as d.RenderNode[]) {
+    const contentRefNode = (Array.from(host.childNodes) as d.RenderNode[]).find((ref) => ref['s-cr']);
+    const childNodeArray = Array.from(parentElm.childNodes) as d.RenderNode[];
+
+    // If we have a content ref, we need to invert the order of the nodes we're relocating
+    // to preserve the correct order of elements in the DOM on future relocations
+    for (const childNode of contentRefNode ? childNodeArray.reverse() : childNodeArray) {
       // Only relocate nodes that were slotted in
       if (childNode['s-sh'] != null) {
-        host.insertBefore(childNode, null);
+        host.insertBefore(childNode, contentRefNode ?? null);
+
         // Reset so we can correctly move the node around again.
         childNode['s-sh'] = undefined;
-
-        // When putting an element node back in its original location,
-        // we need to reset the `slot` attribute back to the value it originally had
-        // so we can correctly relocate it again in the future
-        if (childNode.nodeType === NODE_TYPE.ElementNode && !!childNode['s-sn']) {
-          childNode.setAttribute('slot', childNode['s-sn']);
-        }
 
         // Need to tell the render pipeline to check to relocate slot content again
         checkSlotRelocate = true;
@@ -226,13 +222,6 @@ const putBackInOriginalLocation = (parentElm: Node, recursive: boolean) => {
 
       // Reset so we can correctly move the node around again.
       childNode['s-sh'] = undefined;
-
-      // When putting an element node back in its original location,
-      // we need to reset the `slot` attribute back to the value it originally had
-      // so we can correctly relocate it again in the future
-      if (childNode.nodeType === NODE_TYPE.ElementNode) {
-        childNode.setAttribute('slot', childNode['s-sn'] ?? '');
-      }
 
       checkSlotRelocate = true;
     }
@@ -655,8 +644,11 @@ export const patch = (oldVNode: d.VNode, newVNode: d.VNode, isInitialRender = fa
     }
 
     if (BUILD.vdomAttribute || BUILD.reflect) {
-      if (BUILD.slot && tag === 'slot') {
-        // minifier will clean this up
+      if (BUILD.slot && tag === 'slot' && !useNativeShadowDom) {
+        if (BUILD.experimentalSlotFixes && oldVNode.$name$ !== newVNode.$name$) {
+          newVNode.$elm$['s-sn'] = newVNode.$name$ || '';
+          relocateToHostRoot(newVNode.$elm$.parentElement);
+        }
       } else {
         // either this is the first render of an element OR it's an update
         // AND we already know it's possible it could have changed
@@ -995,9 +987,10 @@ render() {
   if (BUILD.scoped || BUILD.shadowDom) {
     scopeId = hostElm['s-sc'];
   }
+
+  useNativeShadowDom = supportsShadow && (cmpMeta.$flags$ & CMP_FLAGS.shadowDomEncapsulation) !== 0;
   if (BUILD.slotRelocation) {
     contentRef = hostElm['s-cr'];
-    useNativeShadowDom = supportsShadow && (cmpMeta.$flags$ & CMP_FLAGS.shadowDomEncapsulation) !== 0;
 
     // always reset
     checkSlotFallbackVisibility = false;
@@ -1086,26 +1079,6 @@ render() {
                 nodeToRelocate['s-hn'] = nodeToRelocate['s-ol'].parentNode.nodeName;
               }
 
-              // Handle a use-case where we relocate a slot where
-              // the slot name changes along the way (for instance, a default to a named slot).
-              // In this case, we need to update the relocated node's slot attribute to match
-              // the slot name it is being relocated into.
-              //
-              // There is a very niche use case where we may be relocating a text node. For now,
-              // we ignore anything that is not an element node since non-element nodes cannot have
-              // attributes to specify the slot. We'll deal with this if it becomes a problem... but super edge-case-y
-              if (
-                BUILD.experimentalSlotFixes &&
-                nodeToRelocate.nodeType === NODE_TYPE.ElementNode &&
-                slotRefNode['s-fs'] !== nodeToRelocate.getAttribute('slot')
-              ) {
-                if (!slotRefNode['s-fs']) {
-                  nodeToRelocate.removeAttribute('slot');
-                } else {
-                  nodeToRelocate.setAttribute('slot', slotRefNode['s-fs']);
-                }
-              }
-
               // Add it back to the dom but in its new home
               // If we get to this point and `insertBeforeNode` is `null`, that means
               // we're just going to append the node as the last child of the parent. Passing
@@ -1151,7 +1124,7 @@ render() {
   // Hide any elements that were projected through, but don't have a slot to go to.
   // Only an issue if there were no "slots" rendered. Otherwise, nodes are hidden correctly.
   // This _only_ happens for `scoped` components!
-  if (BUILD.experimentalSlotFixes && cmpMeta.$flags$ & CMP_FLAGS.scopedCssEncapsulation) {
+  if (BUILD.experimentalScopedSlotChanges && cmpMeta.$flags$ & CMP_FLAGS.scopedCssEncapsulation) {
     for (const childNode of rootVnode.$elm$.childNodes) {
       if (childNode['s-hn'] !== hostTagName && !childNode['s-sh']) {
         // Store the initial value of `hidden` so we can reset it later when
@@ -1164,6 +1137,9 @@ render() {
       }
     }
   }
+
+  // Clear the content ref so we don't create a memory leak
+  contentRef = undefined;
 };
 
 // slot comment debug nodes only created with the `--debug` flag