+++ test/karma/test-global-script/stencil.config.ts
@@ -2,15 +2,13 @@ import { Config } from '../../../dist/declarations';
 const { WWW_OUT_DIR } = require('../constants');
 
 export const config: Config = {
-  namespace: 'TestInvisibleDefaultPrehydration',
+  namespace: 'TestGlobalScript',
   tsconfig: 'tsconfig.json',
   outputTargets: [
     {
       type: 'www',
       dir: `../${WWW_OUT_DIR}`,
-      empty: false,
-      indexHtml: 'prehydrated-styles.html',
-      serviceWorker: null,
     },
   ],
+  globalScript: 'src/global.ts',
 };